Minor refactoring to keep development in one file
