+++ addons/io_sketchfab_plugin/resources/logo.png
