+++ addons/io_sketchfab_plugin/resources/placeholder.png
