--- addons/io_sketchfab_plugin/sketchfab/__init__.py
@@ -1,272 +1,0 @@
-"""
-Copyright 2021 Sketchfab
-
-Licensed under the Apache License, Version 2.0 (the "License");
-you may not use this file except in compliance with the License.
-You may obtain a copy of the License at
-
-    https://www.apache.org/licenses/LICENSE-2.0
-
-Unless required by applicable law or agreed to in writing, software
-distributed under the License is distributed on an "AS IS" BASIS,
-WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-See the License for the specific language governing permissions and
-limitations under the License.
-"""
-
-import os
-import bpy
-import json
-import shutil
-import tempfile
-from uuid import UUID
-
-class Config:
-
-    ADDON_NAME = 'io_sketchfab'
-    GITHUB_REPOSITORY_URL = 'https://github.com/sketchfab/blender-plugin'
-    GITHUB_REPOSITORY_API_URL = 'https://api.github.com/repos/sketchfab/blender-plugin'
-    SKETCHFAB_REPORT_URL = 'https://help.sketchfab.com/hc/en-us/requests/new?type=exporters&subject=Blender+Plugin'
-
-    SKETCHFAB_URL = 'https://sketchfab.com'
-    DUMMY_CLIENTID = 'hGC7unF4BHyEB0s7Orz5E1mBd3LluEG0ILBiZvF9'
-    SKETCHFAB_OAUTH = SKETCHFAB_URL + '/oauth2/token/?grant_type=password&client_id=' + DUMMY_CLIENTID
-    SKETCHFAB_API = 'https://api.sketchfab.com'
-    SKETCHFAB_SEARCH = SKETCHFAB_API + '/v3/search'
-    SKETCHFAB_MODEL = SKETCHFAB_API + '/v3/models'
-    SKETCHFAB_ORGS = SKETCHFAB_API + '/v3/orgs'
-    SKETCHFAB_SIGNUP = 'https://sketchfab.com/signup'
-
-    BASE_SEARCH = SKETCHFAB_SEARCH + '?type=models&downloadable=true'
-    DEFAULT_FLAGS = '&staffpicked=true&sort_by=-staffpickedAt'
-    DEFAULT_SEARCH = SKETCHFAB_SEARCH + \
-                     '?type=models&downloadable=true' + DEFAULT_FLAGS
-
-    SKETCHFAB_ME = '{}/v3/me'.format(SKETCHFAB_URL)
-    BASE_SEARCH_OWN_MODELS = SKETCHFAB_ME + '/search?type=models&downloadable=true'
-    PURCHASED_MODELS = SKETCHFAB_ME + "/models/purchases?"
-
-    SKETCHFAB_PLUGIN_VERSION = '{}/releases'.format(GITHUB_REPOSITORY_API_URL)
-
-    # Those will be set during plugin initialization, or upon setting a new cache directory
-    SKETCHFAB_TEMP_DIR = ""
-    SKETCHFAB_THUMB_DIR = ""
-    SKETCHFAB_MODEL_DIR = ""
-
-    SKETCHFAB_CATEGORIES = (('ALL', 'All categories', 'All categories'),
-                            ('animals-pets', 'Animals & Pets', 'Animals and Pets'),
-                            ('architecture', 'Architecture', 'Architecture'),
-                            ('art-abstract', 'Art & Abstract', 'Art & Abstract'),
-                            ('cars-vehicles', 'Cars & vehicles', 'Cars & vehicles'),
-                            ('characters-creatures', 'Characters & Creatures', 'Characters & Creatures'),
-                            ('cultural-heritage-history', 'Cultural Heritage & History', 'Cultural Heritage & History'),
-                            ('electronics-gadgets', 'Electronics & Gadgets', 'Electronics & Gadgets'),
-                            ('fashion-style', 'Fashion & Style', 'Fashion & Style'),
-                            ('food-drink', 'Food & Drink', 'Food & Drink'),
-                            ('furniture-home', 'Furniture & Home', 'Furniture & Home'),
-                            ('music', 'Music', 'Music'),
-                            ('nature-plants', 'Nature & Plants', 'Nature & Plants'),
-                            ('news-politics', 'News & Politics', 'News & Politics'),
-                            ('people', 'People', 'People'),
-                            ('places-travel', 'Places & Travel', 'Places & Travel'),
-                            ('science-technology', 'Science & Technology', 'Science & Technology'),
-                            ('sports-fitness', 'Sports & Fitness', 'Sports & Fitness'),
-                            ('weapons-military', 'Weapons & Military', 'Weapons & Military'))
-
-    SKETCHFAB_FACECOUNT = (('ANY', "All", ""),
-                           ('10K', "Up to 10k", ""),
-                           ('50K', "10k to 50k", ""),
-                           ('100K', "50k to 100k", ""),
-                           ('250K', "100k to 250k", ""),
-                           ('250KP', "250k +", ""))
-
-    SKETCHFAB_SORT_BY = (('RELEVANCE', "Relevance", ""),
-                         ('LIKES', "Likes", ""),
-                         ('VIEWS', "Views", ""),
-                         ('RECENT', "Recent", ""))
-
-    SKETCHFAB_SEARCH_DOMAIN = (('DEFAULT', "All site", "", 0),
-                               ('OWN', "Own Models (PRO)", "", 1),
-                               ('STORE', "Store purchases", "", 2))
-
-    MAX_THUMBNAIL_HEIGHT = 512
-
-    SKETCHFAB_UPLOAD_LIMITS = {
-        "basic" : 100 * 1024 * 1024,
-        "plus": 100 * 1024 * 1024,
-        "pro": 200 * 1024 * 1024,
-        "prem": 200 * 1024 * 1024,
-        "biz": 500 * 1024 * 1024,
-        "ent": 500 * 1024 * 1024
-    }
-
-
-class Utils:
-    def humanify_size(size):
-        suffix = 'B'
-        readable = size
-
-        # Megabyte
-        if size > 1048576:
-            suffix = 'MB'
-            readable = size / 1048576.0
-        # Kilobyte
-        elif size > 1024:
-            suffix = 'KB'
-            readable = size / 1024.0
-
-        readable = round(readable, 2)
-        return '{}{}'.format(readable, suffix)
-
-    def humanify_number(number):
-        suffix = ''
-        readable = number
-
-        if number > 1000000:
-            suffix = 'M'
-            readable = number / 1000000.0
-
-        elif number > 1000:
-            suffix = 'K'
-            readable = number / 1000.0
-
-        readable = round(readable, 2)
-        return '{}{}'.format(readable, suffix)
-
-    def build_download_url(uid, use_org_profile=False, active_org=None):
-        if use_org_profile:
-            return '{}/{}/models/{}/download'.format(Config.SKETCHFAB_ORGS, active_org["uid"], uid)
-        else:
-            return '{}/{}/download'.format(Config.SKETCHFAB_MODEL, uid)
-
-    def thumbnail_file_exists(uid):
-        return os.path.exists(os.path.join(Config.SKETCHFAB_THUMB_DIR, '{}.jpeg'.format(uid)))
-
-    def clean_thumbnail_directory():
-        if not os.path.exists(Config.SKETCHFAB_THUMB_DIR):
-            return
-
-        from os import listdir
-        for file in listdir(Config.SKETCHFAB_THUMB_DIR):
-            os.remove(os.path.join(Config.SKETCHFAB_THUMB_DIR, file))
-
-    def clean_downloaded_model_dir(uid):
-        shutil.rmtree(os.path.join(Config.SKETCHFAB_MODEL_DIR, uid))
-
-    def get_thumbnail_url(thumbnails_json):
-        best_height = 0
-        best_thumbnail = None
-        for image in thumbnails_json['images']:
-            if image['height'] <= Config.MAX_THUMBNAIL_HEIGHT and image['height'] > best_height:
-                best_height = image['height']
-                best_thumbnail = image['url']
-
-        return best_thumbnail
-
-    def make_model_name(gltf_data):
-        if 'title' in gltf_data.asset.extras:
-            return gltf_data.asset.extras['title']
-
-        return 'GLTFModel'
-
-    def setup_plugin():
-        if not os.path.exists(Config.SKETCHFAB_THUMB_DIR):
-            os.makedirs(Config.SKETCHFAB_THUMB_DIR)
-
-    def get_uid_from_thumbnail_url(thumbnail_url):
-        return thumbnail_url.split('/')[4]
-
-    def get_uid_from_model_url(model_url, use_org_profile=False):
-        return model_url.split('/')[7] if use_org_profile else model_url.split('/')[5]
-
-    def get_uid_from_download_url(model_url):
-        return model_url.split('/')[6]
-
-    def clean_node_hierarchy(objects, root_name):
-        """
-        Removes the useless nodes in a hierarchy
-        TODO: Keep the transform (might impact Yup/Zup)
-        """
-        # Find the parent object
-        root = None
-        for object in objects:
-            if object.parent is None:
-                root = object
-        if root is None:
-            return None
-
-        # Go down its hierarchy until one child has multiple children, or a single mesh
-        # Keep the name while deleting objects in the hierarchy
-        diverges = False
-        while diverges==False:
-            children = root.children
-            if children is not None:
-
-                if len(children)>1:
-                    diverges = True
-                    root.name = root_name
-
-                if len(children)==1:
-                    if children[0].type != "EMPTY":
-                        diverges = True
-                        root.name = root_name
-                        if children[0].type == "MESH": # should always be the case
-                            matrixcopy = children[0].matrix_world.copy()
-                            children[0].parent = None
-                            children[0].matrix_world = matrixcopy
-                            bpy.data.objects.remove(root)
-                            children[0].name = root_name
-                            root = children[0]
-
-                    elif children[0].type == "EMPTY":
-                        diverges = False
-                        matrixcopy = children[0].matrix_world.copy()
-                        children[0].parent = None
-                        children[0].matrix_world = matrixcopy
-                        bpy.data.objects.remove(root)
-                        root = children[0]
-            else:
-                break
-
-        # Select the root Empty node
-        root.select_set(True)
-
-    def is_valid_uuid(uuid_to_test, version=4):
-        try:
-            uuid_obj = UUID(hex=uuid_to_test, version=version)
-            return True
-        except ValueError:
-            return False
-
-class Cache:
-    SKETCHFAB_CACHE_FILE = os.path.join(
-        bpy.utils.user_resource("SCRIPTS", path="sketchfab_cache", create=True),
-        ".cache"
-    ) # Use a user path to avoid permission-related errors
-
-    def read():
-        if not os.path.exists(Cache.SKETCHFAB_CACHE_FILE):
-            return {}
-
-        with open(Cache.SKETCHFAB_CACHE_FILE, 'rb') as f:
-            data = f.read().decode('utf-8')
-            return json.loads(data)
-
-    def get_key(key):
-        cache_data = Cache.read()
-        if key in cache_data:
-            return cache_data[key]
-
-    def save_key(key, value):
-        cache_data = Cache.read()
-        cache_data[key] = value
-        with open(Cache.SKETCHFAB_CACHE_FILE, 'wb+') as f:
-            f.write(json.dumps(cache_data).encode('utf-8'))
-
-    def delete_key(key):
-        cache_data = Cache.read()
-        if key in cache_data:
-            del cache_data[key]
-
-        with open(Cache.SKETCHFAB_CACHE_FILE, 'wb+') as f:
-            f.write(json.dumps(cache_data).encode('utf-8'))
